examples/vision/requirements.txt
azure-ai-vision-imageanalysis
azure-core
python-dotenv